Add ParseDateTime UDF. Address review
